% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEmInR_det.R
\name{calc.Eall}
\alias{calc.Eall}
\title{Internal function. Calculate total exposed when m>1 in a SEmInR model.}
\usage{
calc.Eall(dat)
}
\arguments{
\item{dat}{Dataframe of the simulation time series.}
}
\value{
Dataframe with additional column.
}
\description{
Internal function. Calculate total exposed when m>1 in a SEmInR model.
}
