% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI.R
\name{GI.bck.discrete}
\alias{GI.bck.discrete}
\title{Internal function. Calculate the _backward_ generation interval distribution with discrete times.}
\usage{
GI.bck.discrete(time.t, g, GI_span, I)
}
\arguments{
\item{time.t}{Integer. Calendar time.}

\item{g}{Numerical vector. Intrinsic generation interval distribution.}

\item{GI_span}{Maximum value for the intrinsic GI.}

\item{I}{Numerical vector. Prevalence time series.}
}
\value{
Density of the backward genration interval distribution.
}
\description{
Internal function. Calculate the _backward_ generation interval distribution with discrete times.
}
