% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI.R
\name{GI.bck}
\alias{GI.bck}
\title{Internal function. Calculate the _backward_ generation interval distribution.}
\usage{
GI.bck(tau, t, df)
}
\arguments{
\item{tau}{Integer. The _index_ of the simulation time (not the time value.)}

\item{t}{Integer. The _index_ of the simulation time (not the time value.)}

\item{df}{Dataframe of the epidemic simulation time series.}
}
\value{
Density of the backward genration interval distribution.
}
\description{
Internal function. Calculate the _backward_ generation interval distribution.
}
