% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESuDe_FCT.R
\name{gi.distrib}
\alias{gi.distrib}
\title{Internal function. Discrete generation interval distribution.}
\usage{
gi.distrib(type, mean, var, span, x)
}
\arguments{
\item{type}{String. Name of the distribution family for the GI. Implemented: \code{'pois'} or  \code{'nbinom'}}

\item{mean}{Numeric. Mean of the GI}

\item{var}{Numeric. Variance of the GI}

\item{span}{Numeric. Maximum value for the GI}

\item{x}{Numeric. Values where the GI must be calculated.}
}
\value{
GI density distribution.
}
\description{
Internal function. Discrete generation interval distribution.
}
