% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI.R
\name{GI.fwd.discrete}
\alias{GI.fwd.discrete}
\title{Internal function. Calculate the _forward_ generation interval distribution with discrete times.}
\usage{
GI.fwd.discrete(time.s, g, GI_span, S)
}
\arguments{
\item{time.s}{Integer. Calendar time.}

\item{g}{Numerical vector. Intrinsic generation interval distribution.}

\item{GI_span}{Maximum value for the intrinsic GI.}

\item{S}{Numerical vector. Susceptible time series.}
}
\value{
Density of the forward genration interval distribution.
}
\description{
Internal function. Calculate the _forward_ generation interval distribution with discrete times.
}
