% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI.R
\name{GI.fwd}
\alias{GI.fwd}
\title{Internal function. Calculate the _forward_ generation interval distribution.}
\usage{
GI.fwd(tau, s, df)
}
\arguments{
\item{tau}{Integer. The _index_ of the simulation time (not the time value.)}

\item{s}{Integer. The _index_ of the simulation time (not the time value.)}

\item{df}{Dataframe of the epidemic simulation time series.}
}
\value{
Density of the forward genration interval distribution.
}
\description{
Internal function. Calculate the _forward_ generation interval distribution.
}
