% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI.R
\name{GI.intrinsic}
\alias{GI.intrinsic}
\title{Internal function. Calculate intrinsic generation interval distribution.}
\usage{
GI.intrinsic(tau, df)
}
\arguments{
\item{tau}{Integer. The _index_ of the simulation time (not the time value.)}

\item{df}{Dataframe of the epidemic simulation time series.}
}
\value{
Density of the intrinsic genration interval distribution.
}
\description{
Internal function. Calculate intrinsic generation interval distribution.
}
