% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI.R
\name{GI.seminr}
\alias{GI.seminr}
\title{Calculates the intrinsic, forward and backward generation interval distributions for SEmInR epidemic model. 
The SEmInR model is an epidemiological model that compartments individuals by stages of infections: S for susceptible, E for exposed (infected but not yet infectious), I infectious and R for individuals removed from the model (following recovery and immunity, deaths, etc.). In order to have realistic residency distributions in compartments E and I, these compartments are artificially sub-divided into m for E, and n for I. 
It is not straightforward to calculate the generation interval distributions for a SEmInR model. This function numericaly solves the three generation interval distributions (intrinsic, backward and forward ones, see [1] for theoretical framework).}
\usage{
GI.seminr(latent_mean, infectious_mean, R0, nE, nI, cal.times.fwdbck, horizon,
  dt = 0.1, I.init = 1e-05)
}
\arguments{
\item{latent_mean}{Numeric. The mean duration of the latent period (in 'E' compartments)}

\item{infectious_mean}{Numeric. The mean duration of the infectious period (in 'I' compartments)}

\item{R0}{Numeric. The basic reproduction number}

\item{nE}{Integer Number of E compatments}

\item{nI}{Integer. Number of I compatments}

\item{cal.times.fwdbck}{Numeric vector. Times where the backward and forward generation interval distributions will be calculated}

\item{horizon}{Numeric.  Time until which the simultion is run. User must make sure it is beyond the end of the epidemic, else the calculation of generation interval distributions may be wrong!}

\item{I.init}{Integer. Initial proportion of infectious individuals. Default value at 1}

\item{pop_size}{Integer. Constant population size. Default value at 1E4}
}
\value{
Returns a named list.
\itemize{
\item{intrinsic:} {Intrinsic generation interval distribution. List composed of two elements: tsi, the time since infection vector and density the associated vector of densities. }
\item{fwd:} {Forward generation interval distribution. List composed of cal.times.fwdbck elements, each one representing the densities of the forward generation interval distribution at the requested calendar times.}
\item{bck:} Backward generation interval distribution. List composed of cal.times.fwdbck elements, each one representing the densities of the backward generation interval distribution at the requested calendar times. 
\item{fwd.mean:} Mean of the forward generation interval distribution. List composed of cal.times.fwdbck elements, each one representing the mean of the forward generation interval distribution at the requested calendar times.
\item{bck.mean:} Mean of the backward generation interval distribution. List composed of cal.times.fwdbck elements, each one representing the mean of the backward generation interval distribution at the requested calendar times.
\item{incidence:} Incidence modeled from the SEmInR model.
\item{susceptible:} Time series of the number of susceptible individuals modeled from the SEmInR model.
}
}
\description{
Calculates the intrinsic, forward and backward generation interval distributions for SEmInR epidemic model. 
The SEmInR model is an epidemiological model that compartments individuals by stages of infections: S for susceptible, E for exposed (infected but not yet infectious), I infectious and R for individuals removed from the model (following recovery and immunity, deaths, etc.). In order to have realistic residency distributions in compartments E and I, these compartments are artificially sub-divided into m for E, and n for I. 
It is not straightforward to calculate the generation interval distributions for a SEmInR model. This function numericaly solves the three generation interval distributions (intrinsic, backward and forward ones, see [1] for theoretical framework).
}
\references{
[1] Champredon D, Dushoff J. Intrinsic and realized generation intervals in infectious-disease transmission. Proceedings of the Royal Society B: Biological Sciences 2015; 282: 20152026.
}
