% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{gi_ct_fit_mle2}
\alias{gi_ct_fit_mle2}
\title{Fit R0 and mean generation interval from contact tracing data}
\usage{
gi_ct_fit_mle2(t.obs, gi.obs, model.epi, fxd.prm, start.optim, CI = 0.95,
  do.plot = FALSE)
}
\arguments{
\item{t.obs}{Numeric vector. Time when the backward generation intervals were observed}

\item{gi.obs}{Numeric vector. Backward generation intervals observed.}

\item{model.epi}{String. Epidemic model used. Choice between \code{seminr} or \code{resude}.}

\item{fxd.prm}{List. Parameters of \code{model.epi} that are fixed.}

\item{CI}{Numeric. Confidence interval level. Default = 0.95.}

\item{do.plot}{Boolean. Fitting diagnostic plots. Default = TRUE.}

\item{R0.rng}{Numeric vector. Values of R0 explored to calculate the likelihood surface.}

\item{gimean.rng}{Numeric vector. Values of mean intrinsic GI explored to calculate the likelihood surface.}
}
\description{
Fit R0 and mean generation interval from contact tracing data
}
