% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gittins0.R
\docType{data}
\name{Gittins0}
\alias{Gittins0}
\title{Gittins0}
\format{
Data matrix with columns
}
\usage{
Gittins0
}
\description{
Gittins index table with discount factor 0 for binary outcomes with prior beta(1,1).
Column represents success, and row represnets failure.
}
\examples{
  Gittins0
}
\keyword{datasets}
