% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gittins05.R
\docType{data}
\name{Gittins05}
\alias{Gittins05}
\title{Gittins05}
\format{
Data matrix with columns
}
\usage{
Gittins05
}
\description{
Gittins index table with discount factor 0.5 for binary outcomes with prior beta(1,1).
Column represents success, and row represnets failure.
}
\examples{
  Gittins05
}
\keyword{datasets}
