% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gittins07.R
\docType{data}
\name{Gittins07}
\alias{Gittins07}
\title{Gittins07}
\format{
Data matrix with columns
}
\usage{
Gittins07
}
\description{
Gittins index table with discount factor 0.99 for binary outcomes with prior beta(1,1).
Column represents success, and row represnets failure.
}
\examples{
  Gittins07
}
\keyword{datasets}
