% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gittins099.R
\docType{data}
\name{Gittins099}
\alias{Gittins099}
\title{Gittins099}
\format{
Data matrix with columns
}
\usage{
Gittins099
}
\description{
Gittins index table with discount factor 0.99 for binary outcomes with prior beta(1,1).
Column represents success, and row represnets failure.
}
\examples{
  Gittins099
}
\keyword{datasets}
