% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gittins0995.R
\docType{data}
\name{Gittins0995}
\alias{Gittins0995}
\title{Gittins0995}
\format{
Data matrix with columns
}
\usage{
Gittins0995
}
\description{
Gittins index table with discount factor 0.995 for binary outcomes with prior beta(1,1).
Column represents success, and row represnets failure.
}
\examples{
  Gittins0995
}
\keyword{datasets}
