% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{nllk}
\alias{nllk}
\title{Negative Log-Likelihood given R0 and mean generation interval
assuming Poisson observation error from contact tracing data}
\usage{
nllk(R0, gimean, t.obs, gi.obs, model.epi, fxd.prm)
}
\arguments{
\item{t.obs}{Numeric vector. Time when the backward generation intervals were observed}

\item{gi.obs}{Numeric vector. Backward generation intervals observed.}

\item{model.epi}{String. Epidemic model used. Choice between \code{seminr} or \code{resude}.}

\item{fxd.prm}{List. Parameters of \code{model.epi} that are fixed.}
}
\description{
Negative Log-Likelihood given R0 and mean generation interval
assuming Poisson observation error from contact tracing data
}
