% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEmInR_det.R
\name{SEmInR.FX}
\alias{SEmInR.FX}
\title{Internal function. Simulate a SEmInR model.}
\usage{
SEmInR.FX(t, x, parms)
}
\arguments{
\item{t}{Numerical vector. Simulation times.}

\item{x}{Numerical named vector. Initial states.}

\item{parms}{Numerical named vector. Parameters of the SEmInR model}
}
\value{
Dataframe of time series for states and probabilities.
}
\description{
Internal function. Simulate a SEmInR model.
}
